A few miscellaneous cleanups.
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/branch/prod/web-ui/work/src@18120 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ buildtools/target.rb
@@ -409,6 +409,11 @@ class ServletBuilder < Target
 
     args += @jsp_list.to_a
 
+    if @jsp_list.empty?
+      debug( "Empty JSP file list." )
+      return
+    end
+
     JavaCompiler.run(cp, "org.apache.jasper.JspC", *args)
 
     FileList["#{@destRoot}/**/*.java"].each { |f| FileUtils.rm(f) }
